Struct owning_bytes::OwningByteBuf

pub struct OwningByteBuf<T> {
    // some fields omitted
}

A wrapper around an array of bytes and an object T that references those bytes.

Methods

impl<T> OwningByteBuf<T>

fn from_vec<'a, F>(buf: Vec<u8>, f: F) -> OwningByteBuf<T> where F: FnOnce(&'a [u8]) -> T

Creates an OwningByteBuf from a vector and a constructing function

fn from_vec_res<'a, F, E>(buf: Vec<u8>, f: F) -> Result<OwningByteBuf<T>, (E, Vec<u8>)> where F: FnOnce(&'a [u8]) -> Result<T, E>

Creates an OwningByteBuf from a vector and a constructing function that may fail

use std::str;
use owning_bytes::OwningByteBuf;

let vec = b"Hello World".to_vec();
let string = OwningByteBuf::from_vec_res(vec, str::from_utf8).unwrap();
assert_eq!(*string.get(), "Hello World");

fn from_box<'a, F>(buf: Box<[u8]>, f: F) -> OwningByteBuf<T> where F: FnOnce(&'a [u8]) -> T

Creates an OwningByteBuf from a boxed slice and a constructing function

fn from_box_res<'a, F, E>(buf: Box<[u8]>, f: F) -> Result<OwningByteBuf<T>, (E, Box<[u8]>)> where F: FnOnce(&'a [u8]) -> Result<T, E>

Creates an OwningByteBuf from a boxed slice and a constructing function that may fail

fn get(&self) -> &T

Returns a reference to the wrapped type.

fn into_vec(self) -> Vec<u8>

Drops the wrapped type and returns the underlying buffer back as a Vec.